Lint, refactor code, fix small bugs
Linted code for consistency and moved enemy to its own file so our index
is a little more manageable.

Fixed a few bugs relating to player controls too

Note: I also updated webpack to webpack 2, so you'll want to run `npm
install` again. Sorry :'(
