+++ README.md
@@ -9,11 +9,14 @@ If you're not familiar, this project uses Node.js. You can download it [here](ht
 1. From your terminal / command prompt, run `npm install`
 2. You can run `npm start` to start the server now
 3. Open `127.0.0.1:8080` in your browser to see the game
-> `localhost` will not work - it is important to use `127.0.0.1`
-4. Any changes you make will auto-refresh the appropriate components on the page. Be aware this may refresh the entire page if you changed the JavaScript core to the game. 
+
+  > `localhost` will not work - it is important to use `127.0.0.1`
+
+4. Any changes you make will auto-refresh the appropriate components on the page. Be aware this may refresh the entire page if you changed the JavaScript core to the game
+5. Before you commit, be sure to run `npm run lint` to lint your code
 
 
-# How To Play 
-Click the Mouse to fire at enemies
+# How To Play
 
-use U, D, R, L keys to move. 
+- Click the Mouse to fire at enemies
+- use your arrow keys to move
